+++ factor/process.py
@@ -695,6 +695,7 @@ def _set_up_directions(parset, bands, dry_run=False, test_run=False,
         direction.flag_abstime = parset['flag_abstime']
         direction.flag_baseline = parset['flag_baseline']
         direction.flag_freqrange = parset['flag_freqrange']
+        direction.flag_expr = parset['flag_expr']
 
         # Reset state if specified
         if direction.name in reset_directions: